Decoupled DemoManager and fixed the bug where Quinton wouldn't talk to the player
