+++ oops-all-bards/Assets/Scripts/Dialogue/DialogueManager.cs
@@ -57,7 +57,7 @@ public class DialogueManager : MonoBehaviour
         questUI.SetActive(false);
         nodeIndex = 0;
         Debug.Log("Starting dialogue.");
-        DemoManager.Instance.TogglePlayerControls();
+        GameManager.Instance.TogglePlayerControls();
         Debug.Log("toggle pause in dialogue");
         dialogueIndex = dialogueID;
         Dialogue dialogue = jsonReader.dialogues.GetDialogue(dialogueID);
@@ -148,7 +148,7 @@ public class DialogueManager : MonoBehaviour
         Dialogue dialogue = jsonReader.dialogues.GetDialogue(dialogueIndex);
         dialogue.Exhausted = true;
         ToggleDialogueUI();
-        DemoManager.Instance.TogglePlayerControls();
+        GameManager.Instance.TogglePlayerControls();
 
     }
