+++ oops-all-bards/Assets/Scripts/GameSystems.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4c98b883948373b43adfafd868b51e37
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
