+++ oops-all-bards/Assets/Scripts/Party/PartyManager.cs
@@ -65,7 +65,7 @@ public class PartyManager : MonoBehaviour
     public void TogglePartyUI()
     {
         partyUI.SetActive(!partyUI.activeSelf);
-        DemoManager.Instance.TogglePlayerControls();
+        GameManager.Instance.TogglePlayerControls();
     }
 
     private void RenderPartyUI()
