+++ oops-all-bards/Assets/Scripts/SceneManagers/VictoryManager.cs
@@ -58,7 +58,7 @@ public class VictoryManager : MonoBehaviour
 
     public void IncrementTavernVisits()
     {
-        DemoManager.Instance.IncrementTavernVisits();
+        GameManager.Instance.IncrementTavernVisits();
     }
 
     public void ToggleCombat()
